Simplify, update README
