+++ examples/event-sourcing-diy/README.md
@@ -4,16 +4,6 @@ The ice cream truck example from [The Dev Owl](https://www.youtube.com/@TheDevOw
 
 The version implemented here is the one that the series ends up with in episide 06, plus the queries from the [How to write a Query Handler Part 2](https://youtu.be/przz3vuAf_M?si=hSQF73OK1Yag2uxJ) episode.
 
-I tried to reproduce the domain as presented in the series as closely as possible in TypeScript.
-
-## Differences from the videos
-
-This example implements some details of the command and query handlers in a style that aligns with Aeon's goals and API.  For example:
-
-1. The business logic behaviors implemented in the series are represented as a discriminated union (F# sum type) of commands and a function to produce new events given a command.
-1. It makes use of features like revision tracking and optimistic concurrency that were outside the scope of the YouTube series.
-1. Rather than a menu-drive CLI, there are separate cli scripts for each operation.
-
 ## Getting started
 
 Create the event store in AWS:
@@ -23,6 +13,23 @@ Create the event store in AWS:
    2. [NodeJS >= 22.12.0 or higher](https://nodejs.org/en/download)
 1. `cd examples/event-store-diy`
 1. `npx cdk deploy`
+
+## Similarities to & Differences from the videos
+
+I tried to reproduce the domain as presented in the series as closely as possible in TypeScript.  Also, like the video, queries are answered by projecting events directly from the event store, i.e. read models are not persisted.
+
+The following files correspond directly to the concepts presented in the videos:
+
+1. **domain.ts** - Domain Events
+2. **projection.ts** - Projections
+3. **behavior.ts** - Behavior(s)
+4. **query.ts** - Query handlers
+
+This example implements some details of the command and query handlers in a different style than the videos to align with Aeon's goals and API.  For example:
+
+1. The business logic behaviors implemented in the series are represented as a discriminated union (F# sum type) of commands and a function to produce new events given a command.
+1. It makes use of features like revision tracking and optimistic concurrency that were outside the scope of the YouTube series.
+1. Rather than a menu-driven CLI, there are separate cli scripts for each operation.
 
 ## CLI
