+++ examples/event-sourcing-diy/cli/restock.mts
@@ -4,13 +4,13 @@ import assert from "node:assert"
 import { readFileSync } from "node:fs"
 import { basename, join } from "node:path"
 import { fromConfigString } from "../../../src/eventstore"
-import { handleCommand } from "../behavior/handler"
+import { handleCommand } from "../behavior"
 import { isFlavour, isTruck } from "../domain"
 
 const { eventSourcingDIY } = JSON.parse(readFileSync(join(import.meta.dirname, '../eventstore.json'), 'utf-8'))
 const store = fromConfigString(eventSourcingDIY.eventStoreConfig, new DynamoDBClient({}))
 
-const [flavour, quantityString, truck] = process.argv.slice(2)
+const [truck, flavour, quantityString] = process.argv.slice(2)
 const quantity = Number.parseInt(quantityString)
 assert(isTruck(truck) && isFlavour(flavour) && Number.isInteger(quantity), `Usage: ${basename(process.argv[1])} <flavour> <quantity> <truck>`)
