+++ examples/event-sourcing-diy/cli/sell.mts
@@ -4,13 +4,13 @@ import assert from "node:assert"
 import { readFileSync } from "node:fs"
 import { basename, join } from "node:path"
 import { fromConfigString } from "../../../src/eventstore"
-import { handleCommand } from "../behavior/handler"
+import { handleCommand } from "../behavior"
 import { isFlavour, isTruck } from "../domain"
 
 const { eventSourcingDIY } = JSON.parse(readFileSync(join(import.meta.dirname, '../eventstore.json'), 'utf-8'))
 const store = fromConfigString(eventSourcingDIY.eventStoreConfig, new DynamoDBClient({}))
 
-const [flavour, truck] = process.argv.slice(2)
+const [truck, flavour] = process.argv.slice(2)
 assert(isTruck(truck) && isFlavour(flavour), `Usage: ${basename(process.argv[1])} <flavour> <truck>`)
 
 handleCommand
